flask/requirements.txt
Flask
pydantic
cassandra-driver
logging